requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.27.8
python-dotenv==1.0.0
pydantic==2.0.2
